standardize side-loaded proof interfaces and relocate examples
