+++ src/examples/token-manager.eg.ts
@@ -25,7 +25,7 @@ import { FungibleToken, VKeyMerkleMap } from '../FungibleTokenContract.js';
 import {
   generateDummyDynamicProof,
   SideloadedProof,
-} from '../side-loaded/program.eg.js';
+} from './side-loaded/program.eg.js';
 import {
   BurnConfig,
   BurnDynamicProofConfig,
