+++ src/test/mint.test.ts
@@ -36,7 +36,7 @@ import {
   generateDynamicProof2,
   SideloadedProof,
   program2,
-} from '../side-loaded/program.eg.js';
+} from '../examples/side-loaded/program.eg.js';
 import {
   CONFIG_PROPERTIES,
   PARAMS_PROPERTIES,
